refactor: move extensions to folder
